LPS-56245 Move TrashHandlers to document library service
