+++ modules/apps/document-library/document-library-service/src/com/liferay/document/library/trash/DLBaseTrashHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.trash;
+package com.liferay.document.library.trash;
 
 import com.liferay.portal.InvalidRepositoryException;
 import com.liferay.portal.kernel.exception.PortalException;
