+++ modules/apps/document-library/document-library-service/src/com/liferay/document/library/trash/DLFileEntryTrashHandler.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.documentlibrary.trash;
+package com.liferay.document.library.trash;
 
 import com.liferay.portal.InvalidRepositoryException;
 import com.liferay.portal.kernel.exception.PortalException;
@@ -20,48 +20,69 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.repository.DocumentRepository;
 import com.liferay.portal.kernel.repository.LocalRepository;
+import com.liferay.portal.kernel.repository.Repository;
 import com.liferay.portal.kernel.repository.RepositoryProviderUtil;
 import com.liferay.portal.kernel.repository.capabilities.TrashCapability;
-import com.liferay.portal.kernel.repository.model.Folder;
+import com.liferay.portal.kernel.repository.model.FileEntry;
+import com.liferay.portal.kernel.repository.util.RepositoryTrashUtil;
+import com.liferay.portal.kernel.search.Field;
+import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.filter.BooleanFilter;
+import com.liferay.portal.kernel.search.filter.Filter;
 import com.liferay.portal.kernel.trash.TrashActionKeys;
-import com.liferay.portal.kernel.trash.TrashRenderer;
+import com.liferay.portal.kernel.trash.TrashHandler;
+import com.liferay.portal.kernel.trash.TrashHandlerRegistryUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.model.ContainerModel;
+import com.liferay.portal.model.TrashedModel;
 import com.liferay.portal.security.permission.ActionKeys;
 import com.liferay.portal.security.permission.PermissionChecker;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portlet.documentlibrary.asset.DLFolderAssetRenderer;
 import com.liferay.portlet.documentlibrary.model.DLFileEntry;
+import com.liferay.portlet.documentlibrary.model.DLFileEntryConstants;
+import com.liferay.portlet.documentlibrary.model.DLFileVersion;
 import com.liferay.portlet.documentlibrary.model.DLFolder;
-import com.liferay.portlet.documentlibrary.model.DLFolderConstants;
 import com.liferay.portlet.documentlibrary.service.DLAppLocalServiceUtil;
 import com.liferay.portlet.documentlibrary.service.DLFileEntryLocalServiceUtil;
+import com.liferay.portlet.documentlibrary.service.DLFileVersionLocalServiceUtil;
 import com.liferay.portlet.documentlibrary.service.DLFolderLocalServiceUtil;
+import com.liferay.portlet.documentlibrary.service.permission.DLFileEntryPermission;
 import com.liferay.portlet.documentlibrary.service.permission.DLFolderPermission;
 import com.liferay.portlet.documentlibrary.util.DLUtil;
+import com.liferay.portlet.documentlibrary.util.DLValidatorUtil;
 import com.liferay.portlet.trash.RestoreEntryException;
 import com.liferay.portlet.trash.TrashEntryConstants;
 import com.liferay.portlet.trash.model.TrashEntry;
 
 import javax.portlet.PortletRequest;
 
+import org.osgi.service.component.annotations.Component;
+
 /**
- * Implements trash handling for the folder entity.
+ * Implements trash handling for the file entry entity.
  *
  * @author Alexander Chow
+ * @author Manuel de la Peña
  * @author Zsolt Berentey
  */
-public class DLFolderTrashHandler extends DLBaseTrashHandler {
+@Component(
+	property = {
+		"model.class.name=com.liferay.portlet.documentlibrary.model.DLFileEntry"
+	},
+	service = TrashHandler.class
+)
+public class DLFileEntryTrashHandler extends DLBaseTrashHandler {
 
 	@Override
 	public void checkRestorableEntry(
 			long classPK, long containerModelId, String newName)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
 		checkRestorableEntry(
-			classPK, 0, containerModelId, dlFolder.getName(), newName);
+			classPK, 0, containerModelId, dlFileEntry.getFileName(),
+			dlFileEntry.getTitle(), newName);
 	}
 
 	@Override
@@ -71,6 +92,7 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 
 		checkRestorableEntry(
 			trashEntry.getClassPK(), trashEntry.getEntryId(), containerModelId,
+			trashEntry.getTypeSettingsProperty("fileName"),
 			trashEntry.getTypeSettingsProperty("title"), newName);
 	}
 
@@ -81,28 +103,33 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 		TrashCapability trashCapability = documentRepository.getCapability(
 			TrashCapability.class);
 
-		Folder folder = documentRepository.getFolder(classPK);
-
-		trashCapability.deleteFolder(folder);
+		trashCapability.deleteFileEntry(
+			documentRepository.getFileEntry(classPK));
 	}
 
 	@Override
 	public String getClassName() {
-		return DLFolder.class.getName();
+		return DLFileEntry.class.getName();
 	}
 
 	@Override
-	public String getDeleteMessage() {
-		return "found-in-deleted-folder-x";
+	public Filter getExcludeFilter(SearchContext searchContext) {
+		BooleanFilter excludeBooleanFilter = new BooleanFilter();
+
+		excludeBooleanFilter.addRequiredTerm(
+			Field.ENTRY_CLASS_NAME, DLFileEntryConstants.getClassName());
+		excludeBooleanFilter.addRequiredTerm(Field.HIDDEN, true);
+
+		return excludeBooleanFilter;
 	}
 
 	@Override
 	public ContainerModel getParentContainerModel(long classPK)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		long parentFolderId = dlFolder.getParentFolderId();
+		long parentFolderId = dlFileEntry.getFolderId();
 
 		if (parentFolderId <= 0) {
 			return null;
@@ -112,14 +139,23 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 	}
 
 	@Override
+	public ContainerModel getParentContainerModel(TrashedModel trashedModel)
+		throws PortalException {
+
+		DLFileEntry dlFileEntry = (DLFileEntry)trashedModel;
+
+		return getContainerModel(dlFileEntry.getFolderId());
+	}
+
+	@Override
 	public String getRestoreContainedModelLink(
 			PortletRequest portletRequest, long classPK)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		return DLUtil.getDLFolderControlPanelLink(
-			portletRequest, dlFolder.getFolderId());
+		return DLUtil.getDLFileEntryControlPanelLink(
+			portletRequest, dlFileEntry.getFileEntryId());
 	}
 
 	@Override
@@ -127,44 +163,33 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 			PortletRequest portletRequest, long classPK)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
 		return DLUtil.getDLFolderControlPanelLink(
-			portletRequest, dlFolder.getParentFolderId());
+			portletRequest, dlFileEntry.getFolderId());
 	}
 
 	@Override
 	public String getRestoreMessage(PortletRequest portletRequest, long classPK)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		return DLUtil.getAbsolutePath(
-			portletRequest, dlFolder.getParentFolderId());
-	}
+		DLFolder dlFolder = dlFileEntry.getFolder();
 
-	@Override
-	public String getSubcontainerModelName() {
-		return "folder";
+		return DLUtil.getAbsolutePath(portletRequest, dlFolder.getFolderId());
 	}
 
 	@Override
 	public String getSystemEventClassName() {
-		return DLFolderConstants.getClassName();
+		return DLFileEntryConstants.getClassName();
 	}
 
 	@Override
 	public TrashEntry getTrashEntry(long classPK) throws PortalException {
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		return dlFolder.getTrashEntry();
-	}
-
-	@Override
-	public TrashRenderer getTrashRenderer(long classPK) throws PortalException {
-		Folder folder = DLAppLocalServiceUtil.getFolder(classPK);
-
-		return new DLFolderAssetRenderer(folder);
+		return dlFileEntry.getTrashEntry();
 	}
 
 	@Override
@@ -175,7 +200,7 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 
 		if (trashActionId.equals(TrashActionKeys.MOVE)) {
 			return DLFolderPermission.contains(
-				permissionChecker, groupId, classPK, ActionKeys.ADD_FOLDER);
+				permissionChecker, groupId, classPK, ActionKeys.ADD_DOCUMENT);
 		}
 
 		return super.hasTrashPermission(
@@ -183,16 +208,11 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 	}
 
 	@Override
-	public boolean isContainerModel() {
-		return true;
-	}
-
-	@Override
 	public boolean isInTrash(long classPK) throws PortalException {
 		try {
-			DLFolder dlFolder = getDLFolder(classPK);
+			DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-			return dlFolder.isInTrash();
+			return dlFileEntry.isInTrash();
 		}
 		catch (InvalidRepositoryException ire) {
 			if (_log.isDebugEnabled()) {
@@ -206,9 +226,9 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 	@Override
 	public boolean isInTrashContainer(long classPK) throws PortalException {
 		try {
-			DLFolder dlFolder = getDLFolder(classPK);
+			DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-			return dlFolder.isInTrashContainer();
+			return dlFileEntry.isInTrashContainer();
 		}
 		catch (InvalidRepositoryException ire) {
 			if (_log.isDebugEnabled()) {
@@ -221,17 +241,17 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 
 	@Override
 	public boolean isRestorable(long classPK) throws PortalException {
-		DLFolder dlFolder = fetchDLFolder(classPK);
+		DLFileEntry dlFileEntry = fetchDLFileEntry(classPK);
 
-		if ((dlFolder == null) ||
-			((dlFolder.getParentFolderId() > 0) &&
+		if ((dlFileEntry == null) ||
+			((dlFileEntry.getFolderId() > 0) &&
 			 (DLFolderLocalServiceUtil.fetchFolder(
-				 dlFolder.getParentFolderId()) == null))) {
+				 dlFileEntry.getFolderId()) == null))) {
 
 			return false;
 		}
 
-		return !dlFolder.isInTrashContainer();
+		return !dlFileEntry.isInTrashContainer();
 	}
 
 	@Override
@@ -240,7 +260,7 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 			ServiceContext serviceContext)
 		throws PortalException {
 
-		DLAppLocalServiceUtil.moveFolder(
+		DLAppLocalServiceUtil.moveFileEntry(
 			userId, classPK, containerModelId, serviceContext);
 	}
 
@@ -252,61 +272,85 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 
 		DocumentRepository documentRepository = getDocumentRepository(classPK);
 
-		TrashCapability trashCapability = documentRepository.getCapability(
-			TrashCapability.class);
-
-		Folder folder = documentRepository.getFolder(classPK);
-
-		Folder destinationFolder = null;
-
-		if (containerModelId != DLFolderConstants.DEFAULT_PARENT_FOLDER_ID) {
-			destinationFolder = documentRepository.getFolder(containerModelId);
-		}
-
-		trashCapability.moveFolderFromTrash(
-			userId, folder, destinationFolder, serviceContext);
+		RepositoryTrashUtil.moveFileEntryFromTrash(
+			userId, documentRepository.getRepositoryId(), classPK,
+			containerModelId, serviceContext);
 	}
 
 	@Override
 	public void restoreTrashEntry(long userId, long classPK)
 		throws PortalException {
 
-		DocumentRepository documentRepository = getDocumentRepository(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		TrashCapability trashCapability = documentRepository.getCapability(
-			TrashCapability.class);
+		if ((dlFileEntry.getClassNameId() > 0) &&
+			(dlFileEntry.getClassPK() > 0)) {
 
-		Folder folder = documentRepository.getFolder(classPK);
+			TrashHandler trashHandler =
+				TrashHandlerRegistryUtil.getTrashHandler(
+					dlFileEntry.getClassName());
 
-		trashCapability.restoreFolderFromTrash(userId, folder);
+			trashHandler.restoreRelatedTrashEntry(getClassName(), classPK);
+
+			return;
+		}
+
+		RepositoryTrashUtil.restoreFileEntryFromTrash(
+			userId, dlFileEntry.getRepositoryId(), classPK);
 	}
 
 	@Override
 	public void updateTitle(long classPK, String name) throws PortalException {
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		dlFolder.setName(name);
+		String fileName = DLUtil.getSanitizedFileName(
+			name, dlFileEntry.getExtension());
 
-		DLFolderLocalServiceUtil.updateDLFolder(dlFolder);
+		dlFileEntry.setFileName(fileName);
+		dlFileEntry.setTitle(name);
+
+		DLFileEntryLocalServiceUtil.updateDLFileEntry(dlFileEntry);
+
+		DLFileVersion dlFileVersion = dlFileEntry.getFileVersion();
+
+		dlFileVersion.setFileName(fileName);
+
+		dlFileVersion.setTitle(name);
+
+		DLFileVersionLocalServiceUtil.updateDLFileVersion(dlFileVersion);
 	}
 
 	protected void checkRestorableEntry(
-			long classPK, long trashEntryId, long containerModelId,
-			String originalTitle, String newName)
+			long classPK, long entryId, long containerModelId,
+			String originalFileName, String originalTitle, String newName)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		if (Validator.isNotNull(newName) &&
+			!DLValidatorUtil.isValidName(newName)) {
+
+			RestoreEntryException ree = new RestoreEntryException(
+				RestoreEntryException.INVALID_NAME);
+
+			ree.setErrorMessage("please-enter-a-valid-name");
+			ree.setTrashEntryId(entryId);
+
+			throw ree;
+		}
+
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
 		if (containerModelId == TrashEntryConstants.DEFAULT_CONTAINER_ID) {
-			containerModelId = DLFolderConstants.DEFAULT_PARENT_FOLDER_ID;
+			containerModelId = dlFileEntry.getFolderId();
 		}
 
 		if (Validator.isNotNull(newName)) {
+			originalFileName = DLUtil.getSanitizedFileName(
+				newName, dlFileEntry.getExtension());
 			originalTitle = newName;
 		}
 
 		DLFolder duplicateDLFolder = DLFolderLocalServiceUtil.fetchFolder(
-			dlFolder.getGroupId(), containerModelId, originalTitle);
+			dlFileEntry.getGroupId(), containerModelId, originalTitle);
 
 		if (duplicateDLFolder != null) {
 			RestoreEntryException ree = new RestoreEntryException(
@@ -314,14 +358,22 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 
 			ree.setDuplicateEntryId(duplicateDLFolder.getFolderId());
 			ree.setOldName(duplicateDLFolder.getName());
-			ree.setTrashEntryId(trashEntryId);
+			ree.setOverridable(false);
+			ree.setTrashEntryId(entryId);
 
 			throw ree;
 		}
 
 		DLFileEntry duplicateDLFileEntry =
 			DLFileEntryLocalServiceUtil.fetchFileEntry(
-				dlFolder.getGroupId(), containerModelId, originalTitle);
+				dlFileEntry.getGroupId(), containerModelId, originalTitle);
+
+		if (duplicateDLFileEntry == null) {
+			duplicateDLFileEntry =
+				DLFileEntryLocalServiceUtil.fetchFileEntryByFileName(
+					dlFileEntry.getGroupId(), containerModelId,
+					originalFileName);
+		}
 
 		if (duplicateDLFileEntry != null) {
 			RestoreEntryException ree = new RestoreEntryException(
@@ -329,11 +381,40 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 
 			ree.setDuplicateEntryId(duplicateDLFileEntry.getFileEntryId());
 			ree.setOldName(duplicateDLFileEntry.getTitle());
-			ree.setOverridable(false);
-			ree.setTrashEntryId(trashEntryId);
+			ree.setTrashEntryId(entryId);
 
 			throw ree;
 		}
+	}
+
+	protected DLFileEntry fetchDLFileEntry(long classPK)
+		throws PortalException {
+
+		Repository repository = RepositoryProviderUtil.getFileEntryRepository(
+			classPK);
+
+		if (!repository.isCapabilityProvided(TrashCapability.class)) {
+			return null;
+		}
+
+		FileEntry fileEntry = repository.getFileEntry(classPK);
+
+		return (DLFileEntry)fileEntry.getModel();
+	}
+
+	protected DLFileEntry getDLFileEntry(long classPK) throws PortalException {
+		Repository repository = RepositoryProviderUtil.getFileEntryRepository(
+			classPK);
+
+		if (!repository.isCapabilityProvided(TrashCapability.class)) {
+			throw new InvalidRepositoryException(
+				"Repository " + repository.getRepositoryId() +
+					" does not support trash operations");
+		}
+
+		FileEntry fileEntry = repository.getFileEntry(classPK);
+
+		return (DLFileEntry)fileEntry.getModel();
 	}
 
 	@Override
@@ -341,11 +422,11 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 		throws PortalException {
 
 		LocalRepository localRepository =
-			RepositoryProviderUtil.getFolderLocalRepository(classPK);
+			RepositoryProviderUtil.getFileEntryLocalRepository(classPK);
 
 		if (!localRepository.isCapabilityProvided(TrashCapability.class)) {
 			throw new InvalidRepositoryException(
-				"LocalRepository " + localRepository.getRepositoryId() +
+				"Repository " + localRepository.getRepositoryId() +
 					" does not support trash operations");
 		}
 
@@ -357,17 +438,19 @@ public class DLFolderTrashHandler extends DLBaseTrashHandler {
 			PermissionChecker permissionChecker, long classPK, String actionId)
 		throws PortalException {
 
-		DLFolder dlFolder = getDLFolder(classPK);
+		DLFileEntry dlFileEntry = getDLFileEntry(classPK);
 
-		if (dlFolder.isInHiddenFolder() && actionId.equals(ActionKeys.VIEW)) {
+		if (dlFileEntry.isInHiddenFolder() &&
+			actionId.equals(ActionKeys.VIEW)) {
+
 			return false;
 		}
 
-		return DLFolderPermission.contains(
-			permissionChecker, dlFolder, actionId);
+		return DLFileEntryPermission.contains(
+			permissionChecker, classPK, actionId);
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(
-		DLFolderTrashHandler.class);
+		DLFileEntryTrashHandler.class);
 
 }
