+++ portal-test-internal/src/com/liferay/portal/service/test/PortalRegisterTestUtil.java
@@ -24,9 +24,6 @@ import com.liferay.portlet.directory.asset.UserAssetRendererFactory;
 import com.liferay.portlet.directory.workflow.UserWorkflowHandler;
 import com.liferay.portlet.documentlibrary.asset.DLFileEntryAssetRendererFactory;
 import com.liferay.portlet.documentlibrary.asset.DLFolderAssetRendererFactory;
-import com.liferay.portlet.documentlibrary.trash.DLFileEntryTrashHandler;
-import com.liferay.portlet.documentlibrary.trash.DLFileShortcutTrashHandler;
-import com.liferay.portlet.documentlibrary.trash.DLFolderTrashHandler;
 import com.liferay.portlet.documentlibrary.util.DLFileEntryIndexer;
 import com.liferay.portlet.documentlibrary.util.DLFolderIndexer;
 import com.liferay.portlet.documentlibrary.workflow.DLFileEntryWorkflowHandler;
@@ -96,9 +93,6 @@ public class PortalRegisterTestUtil {
 			return;
 		}
 
-		TrashHandlerRegistryUtil.register(new DLFileEntryTrashHandler());
-		TrashHandlerRegistryUtil.register(new DLFileShortcutTrashHandler());
-		TrashHandlerRegistryUtil.register(new DLFolderTrashHandler());
 		TrashHandlerRegistryUtil.register(new MBCategoryTrashHandler());
 		TrashHandlerRegistryUtil.register(new MBMessageTrashHandler());
 		TrashHandlerRegistryUtil.register(new MBThreadTrashHandler());
